Fixes U-net arch, changes transforms for facades
Fixes U-net model according to the authors' code that points to
inconsistency with notation in the paper. Specifically, Ck for U-net is
used as if k was a number of channels, not filters, like defined
earlier.

Additionally, fixes transformations like random jitter or flip that
were applied separately to image and to its mask, which means everything
was randomed separately (e.g. image could be flipped but its mask not).
This switch required writing few transformations from torchvision
and switching facades_loader to pass both image and mask to transform.

Moreover, adds support for automatically saving everything to directory
with Google Drive mounted to it (to not lose anything on Colab)

Seems to be converging to something decent.

(giant commit, because it is _impossible_ to split that due to humongous
 base64 in diffs)